Update src/ 	1.) Modularize navbar dropdown into separate component 	2.) Create utility function to manipulate address to print on connect button
