+++ src/utils/trimAddress.js
@@ -1,0 +1,9 @@
+const trimAddress = (acc) => {
+  let trimmedAccount =
+    acc.substring(0, 6) +
+    "..." +
+    acc.substring(acc.length - 4, acc.length);
+  return trimmedAccount;
+};
+
+export default trimAddress;
